terrainops heightmap scale arguments should be floats

+++ TerrainOperations/dllmain.cpp
@@ -3,11 +3,11 @@
 #include "main/dragomath.h"
 
 ex double terrainops_to_heightmap(float* data, unsigned int* out, double len, double scale) {
-	terrainops::to_heightmap(data, out, BYTES2FLOATS(len), scale);
+	terrainops::to_heightmap(data, out, BYTES2FLOATS(len), (float)scale);
 	return 1.0;
 }
 
 ex double terrainops_from_heightmap(float* data, unsigned int* in, double len, double scale) {
-	terrainops::from_heightmap(data, in, BYTES2FLOATS(len), scale);
+	terrainops::from_heightmap(data, in, BYTES2FLOATS(len), (float)scale);
 	return 1.0;
 }
